--- a/examples/webpack-demo-vanilla-bundle/src/examples/example08.scss
+++ b/examples/webpack-demo-vanilla-bundle/src/examples/example08.scss
@@ -1,0 +1,7 @@
+.flex-container {
+  display: flex;
+}
+
+.flex-container > div {
+  padding-right: 5px;
+}