--- a/packages/common/src/filter-conditions/booleanFilterCondition.ts
+++ b/packages/common/src/filter-conditions/booleanFilterCondition.ts
@@ -1,7 +1,17 @@
+import { SearchTerm } from '../enums/index';
 import { FilterCondition, FilterConditionOption } from './../interfaces/index';
 import { parseBoolean } from '../services/utilities';
 
-export const booleanFilterCondition: FilterCondition = (options: FilterConditionOption) => {
-  const searchTerm = Array.isArray(options.searchTerms) && options.searchTerms[0] || '';
-  return parseBoolean(options.cellValue) === parseBoolean(searchTerm);
+/** Execute filter condition check on each cell */
+export const executeBooleanFilterCondition: FilterCondition = (options: FilterConditionOption, parsedSearchValue: boolean) => {
+  return parseBoolean(options.cellValue) === parseBoolean(parsedSearchValue);
 };
+
+/**
+ * From our search filter value(s), get the parsed value(s).
+ * This is called only once per filter before running the actual filter condition check on each cell
+ */
+export function getFilterParsedBoolean(inputSearchTerms: SearchTerm[] | undefined): boolean {
+  const searchTerm = Array.isArray(inputSearchTerms) && inputSearchTerms[0] || false;
+  return parseBoolean(searchTerm);
+}