--- a/packages/common/src/filter-conditions/collectionSearchFilterCondition.ts
+++ b/packages/common/src/filter-conditions/collectionSearchFilterCondition.ts
@@ -1,7 +1,11 @@
 import { FilterCondition, FilterConditionOption } from '../interfaces/index';
 import { testFilterCondition } from './filterUtilities';
 
-export const collectionSearchFilterCondition: FilterCondition = (options: FilterConditionOption) => {
+/**
+ * Execute filter condition check on each cell.
+ * This is used only by the Select Single/Multiple Filter which uses the "multiple-select.js" 3rd party lib which always provide values as strings
+ */
+export const executeCollectionSearchFilterCondition: FilterCondition = (options: FilterConditionOption) => {
   // multiple-select will always return text, so we should make our cell values text as well
   const cellValue = options.cellValue + '';